chore: prepare for mixed package
