+++ python/p3ui/__init__.py
@@ -1,0 +1,1 @@
+from p3ui_core import *
