+++ python/p3ui/matplotlib/__init__.py
@@ -1,0 +1,23 @@
+from .graphics_context import GraphicsContext
+from .renderer import Renderer
+from .figure_manager import FigureManager
+from .figure_canvas import FigureCanvas
+from matplotlib.figure import Figure
+
+
+def new_figure_manager(num, *args, FigureClass=Figure, **kwargs):
+    """Create a new figure manager instance."""
+    # If a main-level app must be created, this (and
+    # new_figure_manager_given_figure) is the usual place to do it -- see
+    # backend_wx, backend_wxagg and backend_tkagg for examples.  Not all GUIs
+    # require explicit instantiation of a main-level app (e.g., backend_gtk3)
+    # for pylab.
+    thisFig = FigureClass(*args, **kwargs)
+    return new_figure_manager_given_figure(num, thisFig)
+
+
+def new_figure_manager_given_figure(num, figure):
+    """Create a new figure manager instance for the given figure."""
+    canvas = FigureCanvas(figure)
+    manager = FigureManager(canvas, num)
+    return manager
