+++ python/p3ui/matplotlib/figure_canvas.py
@@ -1,0 +1,57 @@
+from matplotlib import _api
+from matplotlib.backend_bases import FigureCanvasBase
+
+from p3ui.matplotlib import Renderer
+
+
+class FigureCanvas(FigureCanvasBase):
+    """
+    The canvas the figure renders into.  Calls the draw and print fig
+    methods, creates the renderers, etc.
+    Note: GUI templates will want to connect events for button presses,
+    mouse movements and key presses to functions that call the base
+    class methods button_press_event, button_release_event,
+    motion_notify_event, key_press_event, and key_release_event.  See the
+    implementations of the interactive backends for examples.
+    Attributes
+    ----------
+    figure : `matplotlib.figure.Figure`
+        A high-level Figure instance
+    """
+
+    def __init__(self, *args, **kwargs):
+        super(FigureCanvas, self).__init__(*args, **kwargs)
+
+    def draw(self):
+        """
+        Draw the figure using the renderer.
+        It is important that this method actually walk the artist tree
+        even if not output is produced because this will trigger
+        deferred work (like computing limits auto-limits and tick
+        values) that users may want access to before saving to disk.
+        """
+        print('draw')
+        renderer = Renderer(self.figure.dpi)
+        self.figure.draw(renderer)
+
+    # You should provide a print_xxx function for every file format
+    # you can write.
+
+    # If the file type is not in the base set of filetypes,
+    # you should add it to the class-scope filetypes dictionary as follows:
+    filetypes = {**FigureCanvasBase.filetypes, 'foo': 'My magic Foo format'}
+
+    @_api.delete_parameter("3.5", "args")
+    def print_foo(self, filename, *args, **kwargs):
+        """
+        Write out format foo.
+        This method is normally called via `.Figure.savefig` and
+        `.FigureCanvasBase.print_figure`, which take care of setting the figure
+        facecolor, edgecolor, and dpi to the desired output values, and will
+        restore them to the original values.  Therefore, `print_foo` does not
+        need to handle these settings.
+        """
+        self.draw()
+
+    def get_default_filetype(self):
+        return 'foo'
