+++ python/p3ui/matplotlib/graphics_context.py
@@ -1,0 +1,7 @@
+from matplotlib.backend_bases import GraphicsContextBase
+
+
+class GraphicsContext(GraphicsContextBase):
+
+    def __init__(self):
+        super().__init__()
