+++ python/plot/gradient_chart.py
@@ -1,5 +1,5 @@
 from p3ui import *
-from p3.matplotlib.matplotlib_surface import MatplotlibSurface
+from p3ui.matplotlib.surface import Surface
 import matplotlib.pyplot as plt
 import numpy as np
 
@@ -23,7 +23,7 @@ def gradient_bar(ax, x, y, width=0.5, bottom=0):
                        cmap=plt.cm.Blues_r, cmap_range=(0, 0.8))
 
 
-class GradientChart(MatplotlibSurface):
+class GradientChart(Surface):
     # https://matplotlib.org/stable/gallery/lines_bars_and_markers/bar_stacked.html#sphx-glr-gallery-lines-bars-and-markers-bar-stacked-py
 
     def __init__(self, dpi):
